In chooser lists, treat deleteToBeginningOfLine: as delete:
This is only if there is a responder for the delete: action and is done to allow ⌘⌫ to be used to delete items in the list (like favorites).

The implementation is not ideal as we make assumptions about what action the ⌘⌫ key is bound to and perhaps it’s also a gotcha that ⌘⌫ in a text field will delete something else (though the list is also drawn as having focus).