Remember open files per (project) folder
This means that if one does “mate «folder»”, open a few files as tabs, and then close the window, next time the same folder is opened, TextMate will restore the tabs.

Presently state is remembered (when closing a window) either if the file browser is visible or if more than one document is open in the window.