Set explicit target for status bar menu items
These items were relying on “first responder” which only works when focus is in the document view (since that’s the object that implements the actions).

This would manifest itself by showing the items disabled when focus was in the file browser, and for the grammar pop-up, the selected item (after opening the menu) would not be reverted back to the actual selected grammar in the document view.

Closes #1213