Indicate that find window is not restorable
Since we are using it with an NSDocument, it defaults to use NSDocumentController as restoration class, which gives a warning during launch, if the find window was open during termination, and the conditions are so that the system wants to restore windows from last session.

Also indicate that we do not want to save drafts, unsure if this has any effect.