Variables set via requiredCommands are now shell escaped
Additionally we expect that any variable the user has set is properly shell escaped. Though this is only relevant if the value contains spaces or escape characters.

Currently bundle commands has not been updated to assume that the variables are already shell escaped, so this commit can cause issues with existing bundles.