Get the key window when performing bundle item actions
OakCommitWindow is an NSPanel and therefore cannot become the main window, so bundle actions could not be performed via the bundle menu. Also, we avoid casting it explicitly as a instance of DocumentController for clarity.

This change should be safe since other NSPanels (e.g., Find) do not respond to performBundleItem and it was being handle as a generic id anyways. E.g, bundle actions performed via the menu behave as expected for the BundleEditor, which responds to performBundleItem when the OakTextView is in focus for that window.