Notify VoiceOver user of results count directly
The button for showing results count in find dialog is meant to quickly
provide the user just with the information about number of search results.
Unlike with full search, where the resulting information is very large and
has to be presented in a table, this information the user is requesting
is very short ("%d occurences of \"%s\"") so can be, for better VoiceOver
user experience, presented directly to the user without requiring them to
leave the Show Results Count button. Without this addition, in order to learn
this short information, VoiceOver user would have to move to the status string
below the results table after pressing the button to learn this short
information.

Due to technical reasons, the implementation is not strict:

- it requires the user to stay on the Show Results Count button until
    the results count gets displayed (issue for searches longer than
    1-2 seconds where user might move away from the button), otherwise the
    user is not notified of the result and has to move to the status string
    to learn it
- any changes to the status string not related to results count (e.g. if
    the user has launched a full-fledged search reporting each directory
    searched) are being reported while the user is on the Show Results Count
    button (which should happen really seldom in such a case - Voiceover user
    would probably not "park" their VoiceOver cursor on this button when doing
    a full fledged search)