Show last commit message in the Previous Commit Messages menu
This was previously set to a pop-up menu, hence the need for a placeholder item to set the menu title. However, for pull-down menus, the first item is used as the title and remains fixed.

This difference in behavior caused the last used commit message (first item) not to be listed in the menu.