Do not abort application shutdown when we need to save documents
This was previously necessary because we do threaded saving and relied on our run-loop source to be notified when saving completed, but as application termination (presumably) causes TextMate to run in a special run loop mode, we would not receive this notification.

After switching to dispatch queues there no longer is a problem (at least on OS X 10.9.2).

Closes #1030