Update to support thumbnail generation
QLThumbnail generation is needed. According to Apple docs is more than needed, its basically required more than the full on preview. 

Note that your current code will cause crashes and memory problems as it reads the entire file, which is not a good idea for a quick preview, if the file is 20MB, or 2GB things will get messy. 

The code as I have here has not been compiled, just a few thoughts on what needs to be done. 

--Tom Andersen