Use weak linking for NSAccessibilitySharedFocusElementsAttribute
This is only when building with the 10.10 SDK, so when building with an older SDK we still do a run-time check against the AppKit version and manually initialize the attribute if we are on 10.10 or higher.

This commit has not been tested when built with the 10.10 SDK, nor when running on 10.10 (built either with 10.10 SDK or older).