Add badges to show number of search results (when collapsed)
This is implemented by observing NSOutlineView’s did expand/collapse item notification and then asking for a full reload of the outline view.

None of the granular reload methods would cause the view to be recreated, and there is no “isExpanded” property that the view can bind to.

Given the above, it might be a good idea to introduce an NSTableCellView subclass / view controller for the group view and have this handle it, to avoid the overhead of asking for a full reload each time an item is expanded/collapsed (which is multiplied by number of items when using expand/collapse all).