Document changed on disk would sometimes be marked as modified
We wrap the “update document” in an undo group, but if we already have an undo group at a higher level then our “end group” would be postponed to the root group, which would delay updating the document revision, and thus the revision we use as the disk revision would be wrong.

We workaround this by forcing the undo group to end and thus ignore the one at the higher level, which should be safe.