Do not assume CC/CXX is a single executable
The downside is that we no longer have our .o files depend on the compiler (so we need to do a clean build after upgrading the compiler) and if there is a space in the path to the compiler, it will fail.

The upside is that we can let CC/CXX be a command with arguments, e.g. ‘xcrun clang’.