Allow printing WebViews even when they are not first responder
Starting with 10.9 (I think) it seems that when the WebView switches content view to show a PDF, it will lose first responder status, so pressing ⌘P (to print the PDF) does not work, but require the user to first click the PDF (to make the view first responder).

This commit uses a heuristic to figure out which view should be printed when the first responder is not a printable (web/PDF) view.