Change behavior and implementation of to_s overload for scopes
For a scope context with identical left/right scopes we now return the scope without “decoration” and when the left/right scopes differ, we return them separated by the unit separator character.

The change is motivated by the need to use the result of to_s as a cache key where speed is important.