Set “center tab labels” constraint below “fitting size” priority
This way when asking for a fitting size we get back the minimum size with all but the centering constraint met (and minimized).

In theory the compression resistance for the text field only need to be NSLayoutPriorityFittingSizeCompression+1 but for reasons I do not yet understand, we need to set it higher than NSLayoutPriorityDefaultLow, otherwise the fitting width will occasionally be 64.