Update fonts used for find in folder results
We now use the control content font as this seems to be meant for content shown in tables. We use the standard size (of 12 points) for the headings and set the result rows to use 11 point.

The number of matches is shown with the label font as that is semantically a small font used for placing labels next to things.