Do not relayout tabs to the left of clicked close button
Previously we would still perform layout on the tabs to the left since with tabs having equal width, the layout should result in the same. Now that we compress tabs, the layout can change.

In theory the user could close a tab, have the mouse remain inside the tab bar, and then perform an action that changes the underlying tabs, which would then leave us with tab item views that need layout.