Do not set default settings for Makefile (only source.makefile)
The problem is that if the user changes default settings then these are recorded for the current scope, which has lower precedence than settings set for a glob, so it was effectively impossible for the user to change soft tabs settings for Makefiles (without editing a .tm_properties file).

Also removed the ‘*.ninja’ glob for the same reason.