Change gutter view data source API
This changes behavior slightly as the column width is now decided by the data source rather than based on the widest image.

The gutter view expects the data source to scale images to fit in the available space, since the data source knows about (ideal) line height in the text view and also because scaling the images may lose precision in the image ratio, so scaling it later (if line number font is increased) would give a suboptimal result.